requirements.txt
aiohttp==3.7.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.0.2
psycopg2==2.8.6
pyee==7.0.4
python-decouple==3.4
pytz==2020.4
slackclient==2.9.3
slackeventsapi==2.2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
whitenoise==5.2.0
yarl==1.6.3
